chore: seperated pages/ added joke finder
